feat: seperate components
