+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iSunFA",
-  "version": "0.8.5+34",
+  "version": "0.8.5+35",
   "private": false,
   "scripts": {
     "dev": "next dev",
