+++ src/components/normal_settings/company_settings.tsx
@@ -1,0 +1,47 @@
+import React from 'react';
+import { useTranslation } from 'next-i18next';
+import Image from 'next/image';
+import { Button } from '@/components/button/button';
+
+interface CompanySettingsProps {}
+
+const CompanySettings: React.FC<CompanySettingsProps> = () => {
+  const { t } = useTranslation(['setting', 'common']);
+
+  return (
+    <div className="flex flex-col">
+      <div id="company-setting-section" className="mb-lv-7 flex items-center gap-4">
+        <hr className="block flex-1 border-divider-stroke-lv-4 md:hidden" />
+        <div className="flex items-center gap-2 text-sm text-divider-text-lv-1">
+          <Image src="/icons/asset_management_icon.svg" width={16} height={16} alt="company_icon" />
+          <p>{t('setting:NORMAL.COMPANY_SETTING')}</p>
+        </div>
+        <hr className="flex-1 border-divider-stroke-lv-4" />
+      </div>
+      <Button
+        id="setting-add-company"
+        type="button"
+        variant="linkBorderless"
+        className="mb-lv-7 justify-start p-0"
+      >
+        <p className="flex gap-2">
+          <Image src="/icons/plus.svg" width={16} height={16} alt="notice_icon" />
+          <span>{t('setting:NORMAL.ADD_A_COMPANY')}</span>
+        </p>
+      </Button>
+      <Button
+        id="setting-add-company"
+        type="button"
+        variant="linkBorderless"
+        className="justify-start p-0"
+      >
+        <p className="flex gap-2">
+          <Image src="/icons/notification-text.svg" width={16} height={16} alt="notice_icon" />
+          <span>{t('setting:NORMAL.VIEW_ALL_COMPANIES')}</span>
+        </p>
+      </Button>
+    </div>
+  );
+};
+
+export default CompanySettings;
