+++ src/pages/normal_settings.tsx
@@ -4,7 +4,10 @@ import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 import { ILocale } from '@/interfaces/locale';
 import Head from 'next/head';
 import Layout from '@/components/beta/layout/layout';
-import NormalSettingsBody from '@/components/normal_settings/normal_settings_body';
+import UserSettings from '@/components/normal_settings/user_settings';
+import NoticeSettings from '@/components/normal_settings/notice_settings';
+import CompanySettings from '@/components/normal_settings/company_settings';
+import AccountSettings from '@/components/normal_settings//account_settings';
 
 const NormalSettingsPage: React.FC = () => {
   const { t } = useTranslation(['setting', 'common']);
@@ -19,7 +22,12 @@ const NormalSettingsPage: React.FC = () => {
       </Head>
 
       <Layout isDashboard={false} pageTitle={t('setting:NORMAL.TITLE')}>
-        <NormalSettingsBody />
+        <div className="mx-auto flex w-full max-w-3xl flex-col gap-lv-7 p-lv-4">
+          <UserSettings />
+          <NoticeSettings />
+          <CompanySettings />
+          <AccountSettings />
+        </div>
       </Layout>
     </>
   );
